fix: test utils with random directories/named files
