+++ crates/arey/src/lib.rs
@@ -3,3 +3,6 @@ pub mod ext;
 pub mod svc;
 
 mod log;
+
+#[cfg(test)]
+pub mod test_utils;
