+++ crates/core/src/lib.rs
@@ -8,5 +8,8 @@ pub mod model;
 pub mod session;
 pub mod tools;
 
+#[cfg(test)]
+pub mod test_utils;
+
 pub use crate::assets::get_data_dir;
 pub use crate::provider::llm::get_completion_llm;
